// backend/src/common/response-codes.enum.ts
export enum ResponseCodes {
  GENERIC_OK = '200',
  GENERIC_ACCEPTED = '202',
  GENERIC_BAD_REQUEST = '400',
  GENERIC_NOT_FOUND = '404',
  GENERIC_CONFLICT = '409',
  GENERIC_INTERNAL_SERVER_ERROR = '500',
  GENERIC_CREATED = '201',
  ASSESSMENT_RETRY_AFTER = '429',
  GENERIC_EMPTY_ARRAY = '0000',
  USER_SUCCESS = '0100',
  USER_CREATED = '0101',
  USER_LOGGED_IN = '0102',
  USER_NOT_VERIFIED = '0103',
  USER_NOT_FOUND = '0104',
  USER_NOT_AUTHORIZED = '0105',
  USER_BAD_REQUEST = '0106',
  USER_ALREADY_EXISTS = '0107',
  USER_HAS_PASSWORD = '0108',
  USER_HAS_NO_PASSWORD = '0109',
  USER_EMAIL_LINK_SENT = '0110',
}